requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
jinja2==3.1.6
python-dotenv==1.0.1
discord.py==2.5.2
